connect sprint 3

+++ services/ManageRawMaterialService.js
@@ -1,5 +1,4 @@
 const RawMaterialModel = require('../models/ManageRawMaterialModel');
-const slugify = require('slugify');
 const asyncHandler = require('express-async-handler')
 const Supplier = require('../models/suppliersModel')
 const Manufacturer = require('../models/manufacturersModel');
@@ -49,41 +48,39 @@ exports.getMaterials = asyncHandler(async (req, res) => {
   const userId = req.user._id; // Retrieve user ID from token
   const userType = req.user.userType; // Retrieve user type from token
 
-  const materials = await RawMaterialModel.find({});
+  const materials = await RawMaterialModel.find({ supplierId: userId });
   res.status(200).json({ results: materials.length, data: materials });
 });
 
-// @desc    Get specific Material  by id
-// @route   GET /api/v1/ManageRawMaterial/:id
-// @access  Public
+// @desc Get specific Material  by id
+// @route GET /api/v1/ManageRawMaterial/:id
+// @access Public
 exports.getMaterialById = asyncHandler(async (req, res) => {
-  // const { id } = req.params;
-  // const material = await RawMaterialModel.findById(id);
-  // if (!material) {
-  //   res.status(404).json({ msg: `No material for this id ${id}` });
-  // }
-  // res.status(200).json({ data: material });
   const { id } = req.params; // take id from / :id
+  const userType = req.user.userType;
+  const userId = req.user._id;
 
-  // تحقق مما إذا كان id فارغًا
   if (!id || id.trim() === '') {
     return res.status(400).json({ msg: 'ID is required.' });
   }
 
-  // تحقق من أن الشورت ID يتوافق مع النمط المحدد
   const shortIdPattern = /^[0-9a-z]{8}$/; // Regex for exactly 8 digits
 
-  // تحقق من أن ID مكون من 9 أحرف
   if (id.length !== 8 || !shortIdPattern.test(id)) {
     return res.status(400).json({ msg: `Invalid shortId format: ${id}` });
   }
 
-  // البحث باستخدام shortId
   const material = await RawMaterialModel.findOne({ shortId: id });
 
   // check if the request is null or undefined
   if (!material) {
     return res.status(404).json({ msg: `There is no Material for this id: ${id}` });
+  }
+  const hasAccess =
+    (userType === 'supplier' && material.supplierId.toString() === userId.toString());
+
+  if (!hasAccess) {
+    return res.status(401).json({ msg: 'You do not have permission to get the result.' });
   }
 
   res.status(200).json({ data: material });
@@ -95,6 +92,8 @@ exports.getMaterialById = asyncHandler(async (req, res) => {
 // @access Public
 exports.getMaterialByNameOrId = async (req, res) => {
   const { query } = req.params; // Get query from route parameters
+  const userType = req.user.userType; // Get user type (supplier or manufacturer)
+  const userId = req.user._id; // Get user ID
 
   try {
     // Use findOne to get a single material that matches the name or shortId
@@ -108,6 +107,12 @@ exports.getMaterialByNameOrId = async (req, res) => {
     if (!material) {
       return res.status(404).json({ message: `There is no raw material for this query: ${query}` });
     }
+
+    // Check if the user is the supplier or manufacturer associated with the order
+    if (userType === 'supplier' && material.supplierId.toString() !== userId.toString()) {
+      return res.status(403).json({ msg: 'You do not have permission to get this material.' });
+    }
+
     res.status(200).json({ success: true, data: material });
   } catch (error) {
     console.error('Error fetching material:', error);
@@ -121,13 +126,6 @@ exports.createMaterial = async (req, res) => {
   const userId = req.user._id; // Retrieve user ID from token
   const userType = req.user.userType; // Retrieve user type from token
 
-  console.log('createMaterial', req.body);
-
-  // req.body.slug = slugify(req.body.name);
-  // console.log('createMaterial',req.body);
-
-  // const material = await RawMaterialModel.create(req.body);
-  // res.status(201).json({ data: material });
   try {
     const { name, quantity, description, storageInfo, price, image, materialOption, units } = req.body;
     const material = new RawMaterialModel({
@@ -142,10 +140,10 @@ exports.createMaterial = async (req, res) => {
       units // Add units to the model
     });
 
-    // Update the user's addresses list with the new address ID
+    // Update the Supplier's raw Material list with the new raw Material ID
     await Supplier.findByIdAndUpdate(
       userId,
-      { $addToSet: { rawMaterialList: material._id } }, // استخدم $addToSet لتجنب التكرار
+      { $addToSet: { rawMaterialList: material._id } },
       { new: true }
     );
     await material.save();
@@ -162,17 +160,23 @@ exports.createMaterial = async (req, res) => {
 // @access  Private
 exports.updateMaterial = async (req, res) => {
   const { shortId } = req.body;
-  console.log('updateMaterial', req.body);
+  const userId = req.user._id; // Get user ID
+  const userType = req.user.userType;
 
   try {
-    const material = await RawMaterialModel.findOneAndUpdate(
-      { shortId }, req.body, { new: true }
-    );
-
+    const material = await RawMaterialModel.findOne({ shortId });
     if (!material) {
       res.status(404).json({ msg: `No Material for this id ${shortId}` });
     }
-    res.status(200).json({ success: true, data: material });
+    // Check if the user is associated with the request
+    if (userType === 'supplier' && material.supplierId.toString() !== userId.toString()) {
+      return res.status(403).json({ msg: 'You do not have permission to update this material.' });
+    }
+    const updateMaterial = await RawMaterialModel.findOneAndUpdate(
+      { shortId }, req.body, { new: true }
+    );
+
+    res.status(200).json({ success: true, data: updateMaterial });
 
   } catch (error) {
     console.error('Error updating material:', error);
@@ -184,23 +188,27 @@ exports.updateMaterial = async (req, res) => {
 // @route   DELETE /api/v1/ManageRawMaterial/:id
 // @access  Private
 exports.deleteRawMaterial = async (req, res) => {
-  console.log('deleteRawMaterialbody', req.body);
   const shortId = req.body.data.shortId;
-
-  console.log('deleteRawMaterial', req.body.data.shortId);
+  const userId = req.user._id; // Get user ID
+  const userType = req.user.userType;
 
 
   try {
     // Ensure you're querying by the correct field. Assuming shortId is the field name in your DB:
-    const material = await RawMaterialModel.findOneAndDelete({ shortId });
+    const material = await RawMaterialModel.findOne({ shortId });
 
     if (!material) {
       return res.status(404).json({ message: `No Material found with shortId ${shortId}` });
     }
+    if (userType === 'supplier' && material.supplierId.toString() !== userId.toString()) {
+      return res.status(403).json({ msg: 'Access denied: You do not have permission to delete this material.' });
+    }
+
+    //delete after check permission 
+    await RawMaterialModel.findOneAndDelete({ shortId });
 
     return res.status(200).json({ success: true, msg: 'Material successfully deleted' }); // Changed status to 200
   } catch (error) {
-    console.error('Error deleting material:', error);
     return res.status(500).json({ message: 'Error deleting material', error });
   }
 };
